Add files via upload
The project is a comprehensive data visualization initiative aimed at providing insightful and accessible representations of the impact of the COVID-19 pandemic. Leveraging diverse datasets related to the spread and effects of the virus, this project employs advanced data visualization techniques to communicate key trends, patterns, and critical information to a wide audience. Through interactive charts, graphs, and maps, the project enables users to grasp the geographical distribution of COVID-19 cases, analyze temporal trends, and understand the correlations between various factors. By making use of powerful visualization tools and techniques, the project contributes to enhancing public awareness, facilitating data-driven decision-making, and fostering a better understanding of the evolving landscape of the pandemic.